# swagger.json
{
  "openapi": "3.0.0",
  "components": {
    "securitySchemes": {
      "bearerAuth" : {  
        "type" : "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security":{
    "bearerAuth" : []
  },
  "info": {
    "title": "Proyecto Flutter",
    "description": "API para registrar, listar, y editar Barberos, Clientes, Usuarios y Servicios",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": " http://127.0.0.1:4000"
    },
    {
      "url": "http://10.2.125.234:4000"
    }
  ],
...
  "paths": {
...
    "/obtenerUsuarios": {
      "get": {
        "security":[{
          "bearerAuth" : []
        }],
        "tags":["Usuario"],
        "summary": "Listar todos los Usuarios",
        "description":"Este endpoint permite obtener un listado completo de los usuarios registrados en el sistema.",
        "responses": {
          "200": {
            "description": "Respuesta exitosa que contiene la lista de usuarios.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "usu_id"        : { "type": "integer" },
                      "usu_nombre"    : { "type": "string" },
                      "usu_apellido"  : { "type": "string" },
                      "usu_telefono"  : { "type": "string" },
                      "usu_correo"    : { "type": "string" },
                      "usu_tipo_doc"  : { "type": "string" },
                      "usu_num_doc"   : { "type": "string" }, 
                      "usu_usuario"   : { "type": "string" },
                      "usu_contraseña": { "type": "string" },
                      "usu_estado"    : { "type": "string" },
                      "usu_genero"    : { "type": "string" }
                    },
                    "example": {
                      "usu_id": 7,
                      "usu_nombre": "Mariana",
                      "usu_apellido": "Torres",
                      "usu_telefono": "3114455667",
                      "usu_correo": "marianatorres@example.com",
                      "usu_tipo_doc": "ti",
                      "usu_num_doc": "122334455",
                      "usu_usuario": "marianat",
                      "usu_contraseña": "marianapass321",
                      "usu_estado": "activo",
                      "usu_genero": "Femenino"
                    }
                  }
                }
              }
            }
          }, 
          "404": {
            "description": "error de usuario inexistentes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": { "type": "string" }
                  },
                  "example": {
                    "mensaje": "Ningun registro obtenido."
                  }
                }
              }
            }
          }   
        }
      }
    },
    "/registrarUsuario": {
      "post": {
        "security":[{
          "bearerAuth" : []
        }],
        "tags":["Usuario"],
        "summary": "Registrar un nuevo usuario",
        "description": "Este endpoint permite registrar un nuevo usuario en el sistema. Requiere enviar un objeto JSON con los datos del usuarios como usu_nombre, usu_apellido, usu_telefono, usu_correo, usu_tipo_doc, usu_num_doc, usu_usuario, usu_contraseña, usu_estado y usu_genero.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["'usu_nombre', 'usu_apellido', 'usu_telefono', 'usu_correo', 'usu_tipo_doc', 'usu_num_doc', 'usu_usuario', 'usu_contraseña', 'usu_estado', 'usu_genero'"],
                "properties": {
                  "usu_nombre"    : { "type": "string" },
                  "usu_apellido"  : { "type": "string" },
                  "usu_telefono"  : { "type": "string" },
                  "usu_correo"    : { "type": "string" },
                  "usu_tipo_doc"  : { "type": "string" },
                  "usu_num_doc"   : { "type": "string" }, 
                  "usu_usuario"   : { "type": "string" },
                  "usu_contraseña": { "type": "string" },
                  "usu_estado"    : { "type": "string" },
                  "usu_genero"    : { "type": "string" }
                },
                "example": {
                  "usu_nombre": "Mariana",
                  "usu_apellido": "Torres",
                  "usu_telefono": "3114455667",
                  "usu_correo": "marianatorres@example.com",
                  "usu_tipo_doc": "ti",
                  "usu_num_doc": "122334455",
                  "usu_usuario": "marianat",
                  "usu_contraseña": "marianapass321",
                  "usu_estado": "activo",
                  "usu_genero": "Femenino"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Devuelve un mensaje indicando que el usuario fue registrado correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "properties": {
                    "mensaje": { "type": "string" }
                  },
                  "example": {
                    "mensaje": "Usuario Registrado"
                  }
                }
              }
            }
          },
          "409": {
            "description": "error de entrada duplicada.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": { "type": "string" }
                  },
                  "example": {
                    "mensaje": "ya existe un registro asociado con ese correo"
                  }
                }
              }
            }
          },
          "400": {
            "description": "error de campos faltantes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": { "type": "string" }
                  },
                  "example": {
                    "mensaje": "Faltan campos en la peticion."
                  }
                }
              }
            }
          }
        }
      }
    },
    "/editarUsuario/{usu_id}": {
      "put": {
        "security":[{
          "bearerAuth" : []
        }],
        "tags":["Usuario"],
        "summary": "Editar un usuario existente",
        "description": "Este endpoint permite registrar un nuevo aprendiz en el sistema. Requiere enviar un objeto JSON con los datos del usuarios como usu_nombre, usu_apellido, usu_telefono, usu_correo, usu_tipo_doc, usu_num_doc, usu_usuario, usu_contraseña, usu_estado y usu_genero.",
        "parameters": [
          {
            "name": "usu_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["'usu_nombre', 'usu_apellido', 'usu_telefono', 'usu_correo', 'usu_tipo_doc', 'usu_num_doc', 'usu_usuario', 'usu_contraseña', 'usu_estado', 'usu_genero'"],
                "properties": {
                  "usu_nombre"    : { "type": "string" },
                  "usu_apellido"  : { "type": "string" },
                  "usu_telefono"  : { "type": "string" },
                  "usu_correo"    : { "type": "string" },
                  "usu_tipo_doc"  : { "type": "string" },
                  "usu_num_doc"   : { "type": "string" }, 
                  "usu_usuario"   : { "type": "string" },
                  "usu_contraseña": { "type": "string" },
                  "usu_genero"    : { "type": "string" }
                },
                "example": {
                  "usu_nombre": "Mariana",
                  "usu_apellido": "Torres",
                  "usu_telefono": "3114455667",
                  "usu_correo": "marianatorres@example.com",
                  "usu_tipo_doc": "ti",
                  "usu_num_doc": "122334455",
                  "usu_usuario": "marianat",
                  "usu_contraseña": "marianapass321",
                  "usu_genero": "Femenino"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Devuelve un mensaje indicando que el usuario fue actualizado correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": { "type": "string" }
                  },
                  "example": {
                    "mensaje": "se ha editado el usuario"
                  }
                }
              }
            }
          },
          "404": {
            "description": "error de usuario inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": { "type": "string" }
                  },
                  "example": {
                    "mensaje": "Parece que intentas actualizar un usario que NO existe."
                  }
                }
              }
            }
          },
          "409": {
            "description": "error de entrada duplicada.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": { "type": "string" }
                  },
                  "example": {
                    "mensaje": "ya existe un registro asociado con ese correo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/cambiarEstado/{usu_id}": {
      "put": {
        "security":[{
          "bearerAuth" : []
        }],
        "tags":["Usuario"],
        "summary": "Editar Estado de un usuario",
        "description":"Este endpoint permite editar el estado de un usuario del sistema utilizando su identificador único ID.",
        "parameters": [
          {
            "name": "usu_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["'usu_estado'"],
                "properties": {
                  "usu_estado"    : { "type": "string" }
                },
                "example": {
                  "usu_estado": "Inactivo"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Devuelve un mensaje indicando que el estado del usuario fue actualizado correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": { "type": "string" }
                  },
                  "example": {
                    "mensaje": "Se ha cambiado el estado del usuario"
                  }
                }
              }
            }
          },
          "404": {
            "description": "error de usuario inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": { "type": "string" }
                  },
                  "example": {
                    "mensaje": "Parece que intentas actualizar un usario que NO existe."
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}